tests/wpt/web-platform-tests/old-tests/submission/Infraware/Forms/contents/Forms/fieldset_checkvalidity.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <p>
    <h3>FieldSet_checkValidity</h3>
  </p>
  
  <hr>

  <div id="log"></div>

  <form method="post"
      enctype="application/x-www-form-urlencoded"
      action="" 
      id="input_form">
    <fieldset id="input_field">
     </fieldset>
  </form>
  <script>

    var field = document.getElementById("input_field");

    try
    {
      var ret = field.checkValidity();

      test(function() {
        assert_equals(ret, true, "calling of checkValidity method is failed.");
      });
    }
    catch (e) {
      test(function() {
        assert_unreached("Error is raised.");
      });
    }

  </script>

 </body>
</html>
